move eventsDiscovery component
